This sassy commit is all the rage this summer.
Solves #192 - introduces Sass for CSS generation. Uses Libsassnet to provide the Sassyness.

Switches out ILMerge for Fody/Costura. When handling unmanaged assemblies, this is the best solution. This required some modifications to how MEF was used.
It uses a very specific version of Fody/Costura because of issues with VS15, I think. Please don't update it.